Aorta_UNet_Baseline/requirements.txt
absl-py==2.1.0
alembic==1.13.1
aniso8601==9.0.1
asciitree==0.3.3
astor==0.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.7.0
build==1.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
clear==2.0.0
clearml==1.15.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contextlib2==21.6.0
contourpy==1.2.1
cookiecutter==2.6.0
cucim==23.10.0
cycler==0.12.1
docker==7.0.0
einops==0.7.0
entrypoints==0.4
evalutils==0.4.2
fasteners==0.19
filelock==3.11.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
furl==2.1.3
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.62.1
gunicorn==21.2.0
h5py==3.11.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
imagecodecs==2024.1.1
imageio==2.34.0
importlib_metadata==7.1.0
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
json-tricks==3.17.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lit==18.1.3
lmdb==1.4.1
lpips==0.1.4
Mako==1.3.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mlflow==2.11.3
monai==1.3.0
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
ninja==1.11.1.1
nni==3.0
nptyping==2.5.0
numcodecs==0.12.1
numpy==1.26.4
nvidia-ml-py==12.535.133
onnx==1.16.0
onnxruntime==1.17.3
optuna==3.6.1
orderedmultidict==1.0.1
packaging==23.2
pandas==2.2.2
pathlib2==2.3.7.post1
pillow==10.3.0
pip-tools==7.4.1
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pydicom==2.4.4
PyJWT==2.8.0
pynrrd==1.0.0
pyparsing==3.1.2
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PythonWebHDFS==0.2.3
pytorch-ignite==0.4.11
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
responses==0.25.0
rpds-py==0.18.0
schema==0.7.5
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
SimpleITK==2.3.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.12.1
tqdm==4.66.2
transformers==4.21.3
triton==2.0.0
typeguard==4.1.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
zarr==2.17.2
zipp==3.18.1

batchgenerators==0.25
dicom2nifti
medpy